ppd: Enable sysfs_acpi_monitor by default
If the platform profile is changed outside of tuned then the update
does not get reflected in tuned-adm and gnome-settings.
This is particularly noticeable on Thinkpads where the F8 Mode key
changes the profile but the change is not reflected. It can also be
demonstrated by changing the profile from the command line.

Enable the feature by default. It works well and is useful.

Tested on multiple Thinkpads, but I expect it to be useful to other
vendor platforms too.

Signed-off-by: Mark Pearson <mpearson-lenovo@squebb.ca>