udev-settle: obey udev buffer size and handle possible tracebacks
It now uses TuneD udev events buffer size which is by default 1 MB
instead of the 8 kB default udev buffer size. The buffer size is
configurable in the tuned-main.conf. This should prevent filling
of the buffer even in case of a storm of udev events. It also
contains workaround for the pyudev bug:
https://github.com/pyudev/pyudev/issues/194

Even if the buffer is filled now, the udev queue is no more stuck
and error is logged.

Resolves: RHEL-92637

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>