tuned-ppd: Pin virtual files in memory for inotify
The inodes backing the kernfs virtual files for the attributes we watch
may be dropped by the kernel due to memory exhaustion or explicit user
action (sysctl vm.drop_caches). This causes inotify events to not be
delivered as kernfs_notify() simply drops the event if the inode cannot
be found in the kernel caches.

Fix this by explicitly pinning the attributes virtual files in memory to
ensure we receive inotify events.